@@ -1,5 +1,8 @@
 // ignore_for_file: prefer_typing_uninitialized_variables
 
+import 'dart:async';
+
+import 'package:business_card/language_constants.dart';
 import 'package:business_card/styles/size.dart';
 import 'package:business_card/pages/additional_pages/profile_page.dart';
 import 'package:cached_network_image/cached_network_image.dart';
@@ -7,13 +10,15 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_pagination/firebase_pagination.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_keyboard_visibility/flutter_keyboard_visibility.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
 import '../../styles/colors.dart';
 
 class SearchPage extends StatefulWidget {
-  const SearchPage({super.key});
+  const SearchPage({super.key, required this.isSavedUsers});
 
+  final bool isSavedUsers;
   @override
   State<SearchPage> createState() => _SearchPageState();
 }
@@ -21,10 +26,43 @@ class SearchPage extends StatefulWidget {
 class _SearchPageState extends State<SearchPage> {
   final userEmail = FirebaseAuth.instance.currentUser!.email;
   final TextEditingController _textController = TextEditingController();
+  late final bool isSavedUsers = widget.isSavedUsers;
+  List savedUsers = [];
+  bool isKeyboardOpen = false;
+
+  Future<void> _getSavedUsers() async {
+    final query = await FirebaseFirestore.instance
+        .collection('users')
+        .doc(FirebaseAuth.instance.currentUser!.uid)
+        .get();
+    final data = query.data() as Map<String, dynamic>;
+    if (data['savedUsers'] != null) {
+      savedUsers = data['savedUsers'] as List;
+    }
+  }
+
+  late StreamSubscription<bool> keyboardSubscription;
+  @override
+  void initState() {
+    super.initState();
+    if (isSavedUsers) {
+      _getSavedUsers().whenComplete(() => setState(
+            () {},
+          ));
+    }
+    var keyboardVisibilityController = KeyboardVisibilityController();
+    keyboardSubscription =
+        keyboardVisibilityController.onChange.listen((bool visible) {
+      setState(() {
+        isKeyboardOpen = visible;
+      });
+    });
+  }
 
   @override
   void dispose() {
     _textController.dispose();
+    keyboardSubscription.cancel();
     super.dispose();
   }
 
@@ -51,7 +89,7 @@ class _SearchPageState extends State<SearchPage> {
   Widget build(BuildContext context) {
     return Scaffold(
       resizeToAvoidBottomInset: false,
-      backgroundColor: Colouring.colorLightLightGrey,
+      backgroundColor: Colouring.colorDarkBlue,
       body: SafeArea(
         child: Column(
           children: [
@@ -64,11 +102,13 @@ class _SearchPageState extends State<SearchPage> {
               padding: EdgeInsets.only(
                   left: RelativeSize(context: context)
                       .getScreenWidthPercentage(0.1)),
-              child: const Text(
-                "Find a profile",
-                style: TextStyle(
+              child: Text(
+                isSavedUsers
+                    ? translatedText(context).search_screen_saved_users
+                    : translatedText(context).search_screen_find_profile,
+                style: const TextStyle(
                     fontWeight: FontWeight.w600,
-                    color: Colouring.colorGrey,
+                    color: Colouring.colorAlmostWhite,
                     fontSize: 26),
               ),
             ),
@@ -85,7 +125,7 @@ class _SearchPageState extends State<SearchPage> {
                         .getScreenWidthPercentage(0.05)),
                 child: TextField(
                   controller: _textController,
-                  onChanged: (value) => setState(() {}),
+                  onChanged: (_) => setState(() {}),
                   textAlignVertical: TextAlignVertical.center,
                   decoration: InputDecoration(
                       filled: true,
@@ -95,7 +135,8 @@ class _SearchPageState extends State<SearchPage> {
                       border: OutlineInputBorder(
                           borderSide: BorderSide.none,
                           borderRadius: BorderRadius.circular(30)),
-                      hintText: "Name",
+                      hintText:
+                          translatedText(context).search_screen_hintText_name,
                       contentPadding: EdgeInsets.symmetric(
                           vertical: RelativeSize(context: context)
                               .getScreenWidthPercentage(0.01),
@@ -115,8 +156,21 @@ class _SearchPageState extends State<SearchPage> {
             ),
             Expanded(
               child: Container(
+                padding: EdgeInsets.only(
+                    top: RelativeSize(context: context)
+                        .getScreenHeightPercentage(0.03),
+                    bottom: isKeyboardOpen
+                        ? 0
+                        : RelativeSize(context: context)
+                            .getScreenHeightPercentage(0.12)),
                 decoration: BoxDecoration(
-                    color: Colouring.colorAlmostWhite,
+                    gradient: const LinearGradient(colors: [
+                      Colouring.colorBlueGradient3,
+                      Colouring.colorBlueGradient2
+                    ], stops: [
+                      0.0,
+                      1.0
+                    ], begin: Alignment.topCenter, end: Alignment.bottomCenter),
                     borderRadius: BorderRadius.only(
                         topLeft: Radius.circular(RelativeSize(context: context)
                             .getScreenWidthPercentage(0.13)),
@@ -124,6 +178,7 @@ class _SearchPageState extends State<SearchPage> {
                             .getScreenWidthPercentage(0.13)))),
                 width: double.infinity,
                 child: FirestorePagination(
+                  isLive: true,
                   query: FirebaseFirestore.instance
                       .collection("users")
                       .orderBy("display_name")
@@ -132,25 +187,46 @@ class _SearchPageState extends State<SearchPage> {
                               FirebaseAuth.instance.currentUser!.displayName),
                   itemBuilder: (context, documentSnapshot, index) {
                     final data =
-                        documentSnapshot.data() as Map<dynamic, dynamic>;
-                    print(data);
-                    final user = UserInfo(
-                        data["display_name"], data["email"], data["photoURL"]);
+                        documentSnapshot.data() as Map<String, dynamic>;
+                    if (isSavedUsers) {
+                      if (!savedUsers.contains(documentSnapshot.id)) {
+                        return const SizedBox(
+                          height: 0,
+                          width: 0,
+                        );
+                      }
+                    } else {
+                      if (_textController.text.isEmpty) {
+                        return const SizedBox();
+                      }
+                    }
+                    final user = UserInfo(data["display_name"],
+                        documentSnapshot.id, data["photoURL"]);
                     if (user.displayName
-                            .toLowerCase()
-                            .contains(_textController.text.toLowerCase()) ||
-                        user.email
-                            .toLowerCase()
-                            .contains(_textController.text.toLowerCase())) {
+                        .toLowerCase()
+                        .contains(_textController.text.toLowerCase())) {
                       return GestureDetector(
-                        onTap: () =>
-                            Navigator.of(context).push(_goToProfilePage(user)),
+                        onTap: () async {
+                          await Navigator.of(context)
+                              .push(_goToProfilePage(user))
+                              .then((_) {
+                            _getSavedUsers().whenComplete(() => setState(
+                                  () {},
+                                ));
+                          });
+                        },
                         child: ListTile(
-                          title: Text(user.displayName),
-                          subtitle: Text(user.email),
+                          title: Text(
+                            user.displayName,
+                            style: const TextStyle(
+                                color: Colouring.colorAlmostWhite),
+                          ),
                           leading: user.photoURL == "-1"
                               ? SvgPicture.asset(
                                   'assets/images/icon_profile.svg',
+                                  colorFilter: const ColorFilter.mode(
+                                      Colouring.colorAlmostWhite,
+                                      BlendMode.srcIn),
                                   width: RelativeSize(context: context)
                                       .getScreenWidthPercentage(0.125),
                                 )
@@ -168,10 +244,7 @@ class _SearchPageState extends State<SearchPage> {
                         ),
                       );
                     }
-                    return const SizedBox(
-                      height: 0,
-                      width: 0,
-                    );
+                    return const SizedBox();
                   },
                 ),
               ),
@@ -184,8 +257,8 @@ class _SearchPageState extends State<SearchPage> {
 }
 
 class UserInfo {
-  UserInfo(this.displayName, this.email, this.photoURL);
+  UserInfo(this.displayName, this.uid, this.photoURL);
   late final String displayName;
-  late final String email;
+  late final String uid;
   late final String photoURL;
 }
