@@ -1,10 +1,15 @@
 // ignore_for_file: prefer_const_constructors
+import 'dart:async';
+
+import 'package:business_card/language_constants.dart';
 import 'package:business_card/pages/additional_pages/add_social_media.dart';
+import 'package:business_card/pages/main_pages/settings_page.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:pretty_qr_code/pretty_qr_code.dart';
 
 import 'package:reorderable_grid_view/reorderable_grid_view.dart';
 
@@ -15,17 +20,27 @@ import '../../styles/size.dart';
 import '../../styles/urls.dart';
 
 class EditModePage extends StatefulWidget {
-  const EditModePage(
-      {super.key, required this.userSocialMedia, this.firstConnection = false});
+  const EditModePage({
+    super.key,
+    required this.userSocialMedia,
+    this.firstConnection = false,
+    required this.tabcontroller,
+  });
 
   final dynamic userSocialMedia;
   final bool firstConnection;
+  final TabController tabcontroller;
 
   @override
   State<EditModePage> createState() => _EditModePageState();
 }
 
-class _EditModePageState extends State<EditModePage> {
+class _EditModePageState extends State<EditModePage>
+    with AutomaticKeepAliveClientMixin {
+  @override
+  bool get wantKeepAlive => true;
+  bool isListenerActive = false;
+  late TabController tabController = widget.tabcontroller;
   User? user = FirebaseAuth.instance.currentUser;
   late Map<String, dynamic>? userSocialMedia = widget.userSocialMedia;
   List<String> userSocialMediaAssets = [];
@@ -33,179 +48,294 @@ class _EditModePageState extends State<EditModePage> {
   final _scrollController = ScrollController();
   late bool isReorderable;
   late List<Widget> children;
+  late double containerHeight =
+      RelativeSize(context: context).getScreenHeightPercentage(0.15);
+  Color fadingColor = Colors.transparent;
+  Color profileTabColor = Colouring.colorDarkBlue.withOpacity(0.85);
+  bool isProfileTabOpened = false;
+  bool isVisible = false;
 
   @override
   void initState() {
     super.initState();
+
     isReorderable = false;
     setSocialMedia();
+    tabController.addListener(() {
+      if (isListenerActive && isReorderable) {
+        isListenerActive = false;
+        startTimer();
+      }
+      if (tabController.index != 1) {
+        // Reactivate listener when switching away from index 1
+        isListenerActive = true;
+      }
+    });
+  }
+
+  Timer? timer;
+
+  void startTimer() {
+    if (timer != null && timer!.isActive) {
+      timer!.cancel();
+    }
+    timer = Timer(Duration(seconds: 5), () {
+      if (tabController.index != 1) {
+        saveChanges();
+      }
+      isListenerActive = false; // Reset the flag after timer completes
+    });
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    tabController.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    print(isReorderable);
+    super.build(context);
     return Scaffold(
       backgroundColor: Colors.transparent,
       body: SafeArea(
         child: Scaffold(
           backgroundColor: Colors.transparent,
-          body: Column(
+          body: Stack(
             children: [
-              SizedBox(
-                height: RelativeSize(context: context)
-                    .getScreenHeightPercentage(0.03),
+              Column(
+                children: [
+                  SizedBox(
+                      height: RelativeSize(context: context)
+                              .getScreenWidthPercentage(0.125) +
+                          RelativeSize(context: context)
+                              .getScreenHeightPercentage(0.15)),
+                  getGrid(),
+                  SizedBox(
+                    height: RelativeSize(context: context)
+                        .getScreenHeightPercentage(0.01),
+                  )
+                ],
               ),
-              Center(
-                child: Container(
-                  height: RelativeSize(context: context)
-                      .getScreenHeightPercentage(0.15),
-                  width: RelativeSize(context: context)
-                      .getScreenWidthPercentage(0.875),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(15),
-                    color: Colors.white.withOpacity(0.85),
+              Visibility(
+                visible: isProfileTabOpened,
+                child: GestureDetector(
+                  child: AnimatedContainer(
+                    duration: Duration(milliseconds: 100),
+                    color: fadingColor,
+                    curve: Curves.easeInOut,
                   ),
-                  alignment: Alignment.center,
-                  child: Row(
-                    children: [
-                      Container(
-                        padding: EdgeInsets.only(
-                            left: RelativeSize(context: context)
-                                .getScreenWidthPercentage(0.05)),
-                        child: getProfilePicture(),
-                      ),
-                      SizedBox(
-                        width: RelativeSize(context: context)
-                            .getScreenWidthPercentage(0.05),
-                      ),
-                      Container(
-                        alignment: Alignment.centerLeft,
-                        padding: EdgeInsets.only(
-                            top: RelativeSize(context: context)
-                                .getScreenHeightPercentage(0.055)),
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
+                  onTap: () {
+                    setState(() {
+                      isProfileTabOpened = false;
+                      isVisible = false;
+                      containerHeight = RelativeSize(context: context)
+                          .getScreenHeightPercentage(0.15);
+                      fadingColor = Colors.transparent;
+                      profileTabColor =
+                          Colouring.colorDarkBlue.withOpacity(0.85);
+                    });
+                  },
+                ),
+              ),
+              Container(
+                padding: EdgeInsets.only(
+                    top: RelativeSize(context: context)
+                        .getScreenWidthPercentage(0.0625)),
+                alignment: Alignment.topCenter,
+                child: GestureDetector(
+                  onTap: () {
+                    if (!isReorderable) {
+                      setState(() {
+                        isProfileTabOpened = true;
+                        containerHeight = RelativeSize(context: context)
+                            .getScreenHeightPercentage(0.6);
+                        fadingColor = Colors.black.withOpacity(0.25);
+                        profileTabColor = Colouring.colorDarkBlue;
+                      });
+                    }
+                  },
+                  child: AnimatedContainer(
+                    onEnd: () {
+                      setState(() {
+                        if (isProfileTabOpened) isVisible = true;
+                      });
+                    },
+                    duration: Duration(milliseconds: 200),
+                    curve: Curves.easeInOutCirc,
+                    height: containerHeight,
+                    width: RelativeSize(context: context)
+                        .getScreenWidthPercentage(0.875),
+                    decoration: BoxDecoration(
+                      borderRadius: BorderRadius.circular(15),
+                      color: profileTabColor,
+                    ),
+                    child: Column(
+                      children: [
+                        Row(
                           children: [
-                            Text(
-                              user!.displayName!,
-                              style:
-                                  TextStyle(color: Colors.black, fontSize: 16),
+                            Container(
+                              alignment: Alignment.topCenter,
+                              padding: EdgeInsets.only(
+                                  left: RelativeSize(context: context)
+                                      .getScreenWidthPercentage(0.05),
+                                  top: RelativeSize(context: context)
+                                      .getScreenHeightPercentage(0.021)),
+                              child: getProfilePicture(),
                             ),
                             SizedBox(
-                              height: RelativeSize(context: context)
-                                  .getScreenHeightPercentage(0.0025),
+                              width: RelativeSize(context: context)
+                                  .getScreenWidthPercentage(0.05),
                             ),
-                            Text(
-                              user!.email!,
-                              style: TextStyle(
-                                  color: Color(0x55000000), fontSize: 16),
-                            )
+                            Container(
+                              alignment: Alignment.topCenter,
+                              padding: EdgeInsets.only(
+                                  top: RelativeSize(context: context)
+                                      .getScreenHeightPercentage(0.0225)),
+                              child: Text(
+                                user!.displayName!,
+                                style: TextStyle(
+                                    color: Colouring.colorAlmostWhite,
+                                    fontSize: 16,
+                                    fontWeight: FontWeight.w500),
+                              ),
+                            ),
+                            SizedBox(
+                              width: RelativeSize(context: context)
+                                  .getScreenWidthPercentage(0.03),
+                            ),
+                            Spacer(),
+                            if (userSocialMedia!.isNotEmpty &&
+                                !isReorderable &&
+                                !isProfileTabOpened)
+                              Container(
+                                alignment: Alignment.topRight,
+                                child: PopupMenuButton(
+                                    icon: const Icon(
+                                      Icons.more_vert,
+                                      color: Colouring.colorAlmostWhite,
+                                    ),
+                                    itemBuilder: (context) => [
+                                          PopupMenuItem<int>(
+                                            value: 0,
+                                            padding: EdgeInsets.symmetric(
+                                                horizontal:
+                                                    MediaQuery.of(context)
+                                                            .size
+                                                            .width *
+                                                        0.025),
+                                            child: Row(
+                                              children: [
+                                                Text(translatedText(context)
+                                                    .profile_screen_edit),
+                                                Spacer(),
+                                                Icon(Icons.edit)
+                                              ],
+                                            ),
+                                            onTap: () async {
+                                              setState(() {
+                                                isReorderable = true;
+                                              });
+                                              isListenerActive = true;
+                                            },
+                                          ),
+                                        ]),
+                              )
                           ],
                         ),
-                      ),
-                      SizedBox(
-                        width: RelativeSize(context: context)
-                            .getScreenWidthPercentage(0.03),
-                      ),
-                      Spacer(),
-                      if (userSocialMedia!.isNotEmpty)
-                        Container(
-                          alignment: Alignment.topRight,
-                          child: PopupMenuButton(
-                              icon: const Icon(Icons.more_vert),
-                              itemBuilder: (context) => [
-                                    PopupMenuItem<int>(
-                                      value: 0,
-                                      padding: EdgeInsets.symmetric(
-                                          horizontal: MediaQuery.of(context)
-                                                  .size
-                                                  .width *
-                                              0.025),
-                                      child: Row(
-                                        children: const [
-                                          Text("Edit"),
-                                          Spacer(),
-                                          Icon(Icons.edit)
-                                        ],
-                                      ),
-                                      onTap: () {
-                                        setState(() {
-                                          isReorderable = true;
-                                        });
-                                      },
-                                    ),
-                                  ]),
-                        )
-                    ],
+                        Visibility(
+                          visible: isVisible,
+                          child: SizedBox(
+                            height: RelativeSize(context: context)
+                                .getScreenHeightPercentage(0.03),
+                          ),
+                        ),
+                        Visibility(
+                          visible: isVisible,
+                          child: ElevatedButton(
+                            onPressed: () {
+                              Navigator.of(context)
+                                  .push(_goToSettingsPage())
+                                  .then((value) => setState(() {
+                                        user =
+                                            FirebaseAuth.instance.currentUser!;
+                                      }));
+                            },
+                            style: ButtonStyle(
+                                elevation: MaterialStateProperty.all(0),
+                                backgroundColor: MaterialStateProperty.all(
+                                    Colors.transparent),
+                                overlayColor: MaterialStateProperty.all(
+                                    const Color.fromARGB(255, 0, 19, 31)),
+                                fixedSize: MaterialStateProperty.all(Size(
+                                    double.infinity,
+                                    RelativeSize(context: context)
+                                        .getScreenHeightPercentage(0.06))),
+                                shape: MaterialStateProperty.all(
+                                  RoundedRectangleBorder(
+                                    borderRadius:
+                                        BorderRadius.zero, // No rounded borders
+                                    // Border color with opacity
+                                  ),
+                                )),
+                            child: Padding(
+                              padding: EdgeInsets.only(
+                                  left: RelativeSize(context: context)
+                                      .getScreenWidthPercentage(0.0325),
+                                  right: RelativeSize(context: context)
+                                      .getScreenWidthPercentage(0.075)),
+                              child: Row(
+                                children: [
+                                  SvgPicture.asset(
+                                    'assets/images/settings_icon.svg',
+                                    colorFilter: const ColorFilter.mode(
+                                        Colouring.colorAlmostWhite,
+                                        BlendMode.srcIn),
+                                  ),
+                                  Expanded(
+                                      child: Text(
+                                    translatedText(context)
+                                        .profile_screen_settings,
+                                    style: TextStyle(
+                                        color: Colouring.colorAlmostWhite,
+                                        fontSize: 22,
+                                        fontWeight: FontWeight.w500),
+                                    textAlign: TextAlign.center,
+                                  ))
+                                ],
+                              ),
+                            ),
+                          ),
+                        ),
+                        Visibility(
+                          visible: isVisible,
+                          child: Expanded(
+                            child: Center(
+                              child: SizedBox(
+                                  height: RelativeSize(context: context)
+                                      .getScreenWidthPercentage(0.6),
+                                  width: RelativeSize(context: context)
+                                      .getScreenWidthPercentage(0.6),
+                                  child: PrettyQrView.data(
+                                    decoration: const PrettyQrDecoration(
+                                        shape: PrettyQrSmoothSymbol(
+                                            color: Colouring.colorAlmostWhite)),
+                                    data: user!.uid,
+                                  )),
+                            ),
+                          ),
+                        ),
+                      ],
+                    ),
                   ),
                 ),
               ),
-              SizedBox(
-                  height: RelativeSize(context: context)
-                      .getScreenHeightPercentage(0.025)),
-              getGrid(),
-              SizedBox(
-                height: RelativeSize(context: context)
-                    .getScreenHeightPercentage(0.01),
-              )
             ],
           ),
         ),
       ),
-      //  ])),
     );
   }
-
-  /*Widget getGrid() {
-    setSocialMedia();
-    return Expanded(
-      child: CustomScrollView(
-        slivers: [
-          SliverPadding(
-            padding: EdgeInsets.symmetric(
-                horizontal: RelativeSize(context: context)
-                    .getScreenWidthPercentage(0.075)),
-            sliver: SliverGrid(
-              delegate: SliverChildBuilderDelegate((context, index) {
-                return index != userSocialMediaAssets.length
-                    ? GestureDetector(
-                        onTap: () => goToUserSocialMedia(index),
-                        child: Container(
-                            alignment: Alignment.center,
-                            decoration: BoxDecoration(
-                              borderRadius: BorderRadius.circular(
-                                  RelativeSize(context: context)
-                                      .getScreenWidthPercentage(0.1)),
-                              image: DecorationImage(
-                                  image:
-                                      AssetImage(userSocialMediaAssets[index])),
-                            )),
-                      )
-                    : GestureDetector(
-                        onTap: () => addSocialMedia(),
-                        child: Container(
-                          alignment: Alignment.center,
-                          decoration: BoxDecoration(
-                              borderRadius: BorderRadius.circular(
-                                  RelativeSize(context: context)
-                                      .getScreenWidthPercentage(0.1))),
-                          child: SvgPicture.asset(
-                              'assets/images/add_social_media.svg'),
-                        ),
-                      );
-              }, childCount: userSocialMedia!.values.toList().length + 1),
-              gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-                  crossAxisCount: 2,
-                  mainAxisSpacing: RelativeSize(context: context)
-                      .getScreenWidthPercentage(0.075),
-                  crossAxisSpacing: RelativeSize(context: context)
-                      .getScreenWidthPercentage(0.075)),
-            ),
-          )
-        ],
-      ),
-    );
-  }*/
 
   Widget getGrid() {
     isReorderable ? setEditedChildren() : setChildren();
@@ -213,8 +343,12 @@ class _EditModePageState extends State<EditModePage> {
         child: Stack(children: [
       ListView(children: [
         SizedBox(
-          height:
-              RelativeSize(context: context).getScreenHeightPercentage(0.935),
+          //formula to calculate height of the listview
+          height: (RelativeSize(context: context)
+                      .getScreenWidthPercentage(0.425) *
+                  ((children.length + 1) / 2).ceil()) +
+              (RelativeSize(context: context).getScreenWidthPercentage(0.075) *
+                  ((children.length - 1) / 4).floor()),
           child: ReorderableGridView.count(
             dragEnabled: isReorderable,
             onReorder: (oldIndex, newIndex) {
@@ -262,7 +396,8 @@ class _EditModePageState extends State<EditModePage> {
       ]),
       if (isReorderable)
         Positioned(
-          bottom: 10,
+          bottom:
+              RelativeSize(context: context).getScreenHeightPercentage(0.125),
           left: 20,
           right: 20,
           height:
@@ -275,77 +410,15 @@ class _EditModePageState extends State<EditModePage> {
                   backgroundColor: MaterialStatePropertyAll(Colors.white),
                   overlayColor: MaterialStatePropertyAll(Color(0xFFDFDFDF))),
               onPressed: () async {
-                setState(() {
-                  isReorderable = false;
-                });
-
-                final firestoreDatabase = FirebaseFirestore.instance;
-
-                userSocialMedia!.forEach((key, value) async {
-                  await firestoreDatabase
-                      .collection("users")
-                      .doc(user!.uid)
-                      .set({
-                    "social_media": {
-                      key: {
-                        "position":
-                            userSocialMedia!.keys.toList().indexOf(key) + 1,
-                        "url": value
-                      }
-                    }
-                  }, SetOptions(merge: true));
-                });
+                saveChanges();
               },
               child: Text(
-                "Save changes",
+                translatedText(context).profile_screen_save_changes,
                 style: TextStyle(color: Colouring.colorGrey, fontSize: 20),
               )),
         )
     ]));
   }
-
-  /*
-  SliverPadding(
-            padding: EdgeInsets.symmetric(
-                horizontal: RelativeSize(context: context)
-                    .getScreenWidthPercentage(0.075)),
-            sliver: SliverGrid(
-              delegate: SliverChildBuilderDelegate((context, index) {
-                return index != userSocialMediaAssets.length
-                    ? GestureDetector(
-                        onTap: () => goToUserSocialMedia(index),
-                        child: Container(
-                            alignment: Alignment.center,
-                            decoration: BoxDecoration(
-                              borderRadius: BorderRadius.circular(
-                                  RelativeSize(context: context)
-                                      .getScreenWidthPercentage(0.1)),
-                              image: DecorationImage(
-                                  image:
-                                      AssetImage(userSocialMediaAssets[index])),
-                            )),
-                      )
-                    : GestureDetector(
-                        onTap: () => addSocialMedia(),
-                        child: Container(
-                          alignment: Alignment.center,
-                          decoration: BoxDecoration(
-                              borderRadius: BorderRadius.circular(
-                                  RelativeSize(context: context)
-                                      .getScreenWidthPercentage(0.1))),
-                          child: SvgPicture.asset(
-                              'assets/images/add_social_media.svg'),
-                        ),
-                      );
-              }, childCount: userSocialMedia!.values.toList().length + 1),
-              gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-                  crossAxisCount: 2,
-                  mainAxisSpacing: RelativeSize(context: context)
-                      .getScreenWidthPercentage(0.075),
-                  crossAxisSpacing: RelativeSize(context: context)
-                      .getScreenWidthPercentage(0.075)),
-            ),
-          )*/
 
   //set children list(usma)
   void setChildren() {
@@ -355,7 +428,6 @@ class _EditModePageState extends State<EditModePage> {
               key: ValueKey(index),
               onTap: () => goToUserSocialMedia(index),
               child: Container(
-                  padding: EdgeInsets.all(5),
                   alignment: Alignment.center,
                   decoration: BoxDecoration(
                     borderRadius: BorderRadius.circular(
@@ -369,7 +441,6 @@ class _EditModePageState extends State<EditModePage> {
               key: ValueKey(index),
               onTap: () => addSocialMedia(),
               child: Container(
-                padding: EdgeInsets.all(5),
                 alignment: Alignment.center,
                 decoration: BoxDecoration(
                     borderRadius: BorderRadius.circular(
@@ -379,6 +450,61 @@ class _EditModePageState extends State<EditModePage> {
               ),
             );
     });
+  }
+
+  Route<Object> _goToSettingsPage() {
+    return PageRouteBuilder(
+        pageBuilder: (context, animation, secondaryAnimation) =>
+            const SettingsPage(),
+        transitionsBuilder: (context, animation, secondaryAnimation, child) {
+          const begin = Offset(1, 0);
+          const end = Offset.zero;
+          const curve = Curves.easeInOut;
+
+          final tween =
+              Tween(begin: begin, end: end).chain(CurveTween(curve: curve));
+
+          return SlideTransition(
+            position: animation.drive(tween),
+            child: child,
+          );
+        });
+  }
+
+//saved made changes when reordering and/or deleting widgets
+  void saveChanges() async {
+    setState(() {
+      isReorderable = false;
+    });
+
+    final firestoreDatabase = FirebaseFirestore.instance;
+    Map<String, Map<String, dynamic>> obj = {};
+    for (int i = 0; i < userSocialMedia!.keys.toList().length; i++) {
+      String key = userSocialMedia!.keys.toList()[i];
+      obj.addAll({
+        key: {
+          "position": userSocialMedia!.keys.toList().indexOf(key) + 1,
+          "url": userSocialMedia!.values.toList()[i]
+        }
+      });
+    }
+    await firestoreDatabase
+        .collection("users")
+        .doc(user!.uid)
+        .update({"social_media": obj});
+    /*for (int i = 0; i < userSocialMedia!.keys.toList().length; i++) {
+      String key = userSocialMedia!.keys.toList()[i];
+      await firestoreDatabase.collection("users").doc(user!.uid).update(i != 0
+          ? {
+              "social_media": {
+                key: {
+                  "position": userSocialMedia!.keys.toList().indexOf(key) + 1,
+                  "url": userSocialMedia!.values.toList()[i]
+                }
+              }
+            }
+          : {"social_media": {}});
+    }*/
   }
 
   //set children list(usma) with a delete button
@@ -438,6 +564,7 @@ class _EditModePageState extends State<EditModePage> {
             backgroundImage: CachedNetworkImageProvider(user!.photoURL!),
           )
         : Icon(Icons.account_circle_outlined,
+            color: Colouring.colorAlmostWhite,
             size:
                 RelativeSize(context: context).getScreenWidthPercentage(0.20));
   }
